Version 0 - Initial Design
The first complete version of my project (TechStoreStockInformation) with no feedback or improvements made.

The main features of my project are the SQLite database, which contains all of the information on the stock in the tech shop. For example, the product ID, name, retail price, manufacturing cost, weight, and number of stock. I have linked this to Python, where the user can choose the following queries: Sort by product ID,  retail price or manufacturing cost.