Version 1 - Feedback Received From Three People
Program has been improved with feedback from three different people. It now has a descending and ascending option for each query, as well as giving the user an option to choose to order the products via weight or number of stock. Along with this, I have also made some general improvements to the code.