Version 3 -Feedback from another 3 people
The final version of my project.

A lot has been improved since version 2. I have applied the feedback from the three people, firstly, I added an additional table to the database so the user can get more functionality out of the program. Secondly, I improved the format of the output, it now uses the pretty table module, which is very nice looking and tidy. And finally, I added a password to the program so it is more secure. I also made some general improvements to the rest of the code as well as commenting it to prove that I understand the code.